Safari-611.1.10/LayoutTests/crypto/subtle/hmac-export-key-malformed-parameters.html

<!DOCTYPE html>
<html>
<head>
<script src="../../resources/js-test-pre.js"></script>
<script src="../resources/common.js"></script>
</head>
<body>
<p id="description"></p>
<div id="console"></div>

<script>
description("Test exporting an HMAC key with malformed parameters");

jsTestIsAsync = true;

var extractable = true;

crypto.subtle.generateKey({name: "hmac", hash: "sha-1"}, extractable, ["sign", "verify"]).then(function(result) {
    key = result;

    // Not support format.
    return shouldReject('crypto.subtle.exportKey("spki", key)');
}).then(function() {
    return shouldReject('crypto.subtle.exportKey("pkcs8", key)');
}).then(finishJSTest, finishJSTest);
</script>

<script src="../../resources/js-test-post.js"></script>
</body>
</html>
